ci-requirements.txt
# Dependencies for linting and tests
# Flexible versions avoid --require-hashes mode
pydantic-settings>=2.10,<3.0
prometheus_client>=0.19
pytest>=8.2
pytest-asyncio>=1.1
pytest-cov>=5.0
pytest-mock>=3.14
requests>=2.32
respx[router]>=0.22
httpx>=0.28
tenacity>=9
fastapi>=0.116
uvicorn>=0.30
pyyaml>=6.0
ruff>=0.12
pip-tools>=7.4
setuptools>=80.9
wheel>=0.45
pip>=24.3
